+++ src/extension.ts
@@ -1,48 +1,18 @@
-"use strict";
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
 import * as vscode from "vscode";
-import { PM2, Process } from "./pm2";
+import { registerCommands } from "./contributions/commands";
+import { registerData } from "./contributions/data";
+import { PM2 } from "./model";
 
 let pm2: PM2;
 
-// this method is called when your extension is activated
-// your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
-    pm2 = new PM2(context);
+    const settings = vscode.workspace.getConfiguration("pm2Explorer");
+    pm2 = new PM2(context, settings);
 
-    vscode.window.registerTreeDataProvider("pm2-processes", pm2);
-
-    vscode.commands.registerCommand("pm2.reload", (item: Process) => {
-        pm2.reload(item.process);
-    });
-
-    vscode.commands.registerCommand("pm2.logs", (item: Process) => {
-        pm2.logs(item.process);
-    });
-
-    vscode.commands.registerCommand("pm2.start", (item: Process) => {
-        item.start();
-    });
-
-    vscode.commands.registerCommand("pm2.stop", (item: Process) => {
-        item.stop();
-    });
-
-    vscode.commands.registerCommand("pm2.logsAll", () => {
-        pm2.logs();
-    });
-
-    vscode.commands.registerCommand("pm2.reloadAll", () => {
-        pm2.reloadAll();
-    });
-
-    vscode.commands.registerCommand("pm2.refresh", () => {
-        pm2.refresh();
-    });
+    registerCommands(pm2);
+    registerData(pm2);
 }
 
-// this method is called when your extension is deactivated
 export function deactivate() {
     pm2.dispose();
 }