+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+### 0.0.4
+
+    - Improved process icons
+    - Configurable refresh interval
+    - Added start all and stop all commands
+    - Refactor and re-organisation
+
+### 0.0.3
+
+    - Visual representation of process state
+
 ### 0.0.2
 
     - Fix errors in readme